Desktop/CryptoBits-ML-Model-master/CryptoBits-ML-Model-master/requirements.txt
tensorflow
yfinance
numpy
pymongo
gevent
flask
flask_cors
scikit-learn
pandas
requests
beautifulsoup4
python-dotenv
# pyuwsgi